back-end/requirements.txt
pyswip
python-dotenv
google-genai
fastapi
uvicorn
pytest
pytest-cov
httpx
